feat: Basic Functionality Integration Test
